requirements.txt
aiohttp==3.8.1
aiosignal==1.3.2
async-timeout==4.0.3
attrs==24.3.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==2.1.1
click==8.1.8
cryptography==44.0.0
Deprecated==1.2.16
frozenlist==1.5.0
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.1.0
propcache==0.2.1
pycparser==2.22
PyGithub==2.5.0
Pygments==2.19.1
PyJWT==2.10.1
PyNaCl==1.5.0
requests==2.32.3
rich==13.9.4
shellingham==1.5.4
typer==0.15.1
typing_extensions==4.12.2
urllib3==2.3.0
wrapt==1.17.2
yarl==1.18.3
